requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
openai-whisper==20250625
pydantic-settings==2.10.1
